Use decorator pattern for StudentRiskLevel presentation
+ Move risk level explanation from the model itself to the decorator
