+++ app/views/homerooms/show.html.erb
@@ -224,7 +224,9 @@
   </script>
 
   <% @rows.each do |row| %>
-    <script class="risk-level-tooltip" type="x-tmpl-mustache" data-student-id="<%= row[:student_risk_level]['student_id'] %>">
+    <script class="risk-level-tooltip"
+            type="x-tmpl-mustache"
+            data-student-id="<%= row[:student_risk_level]['student_id'] %>" >
       <div class="tooltip-popup student-risk-level-tooltip">
         <%= row[:student_risk_level]['explanation'].html_safe %>
       </div>
