+++ app/views/students/show.html.erb
@@ -13,8 +13,8 @@
       <h1 id="student-name"><%= @presenter.full_name %></h1>
       <div id="profile-risk-level" class="tooltip-trigger">
         <p>Risk level</p>
-        <div id="risk-pill" class="<%= StudentRiskLevelPresenter.new(@level).css_class_name %>">
-          <h3><%= StudentRiskLevelPresenter.new(@level).level_as_string %></h3>
+        <div id="risk-pill" class="<%= @student_risk_level.css_class_name %>">
+          <h3><%= @student_risk_level.level_as_string %></h3>
         </div>
       </div>
     </div>
@@ -134,7 +134,7 @@
     <div class="tooltip-popup">
       <div class="warning-key">
         <div class="warning-header">
-          Risk level: <%= StudentRiskLevelPresenter.new(@level).level_as_string %>
+          Risk level: <%= @student_risk_level.level_as_string %>
         </div>
         <div class="warning-set">
           <%= @student_risk_level.explanation.html_safe %>
